bytecode: Implement for loop with step range
Add OpStepRange which pushes the result of evaluating the condition of
the loop onto the stack to be evaluated by OpJumpOnFalse. Both the
initialization of the loop variable, and the increment are output as
bytecode. This focuses the complexity in the compiler and follows our
general philosophy of "keeping the vm simple".

Co-authored-by: joshcarp <joshcarp@users.noreply.github.com>