bytecode: Add order to mapVal
Maps cannot be ranged over reliably without an order for the keys. This
change adds a separate slice to the mapVal that keeps an ordered list
of keys. To make filling the order slice easier, the operand for OpMap
is now the length of the map, and not the length multiplied by 2.

Co-authored-by: joshcarp <joshcarp@users.noreply.github.com>