builtins: Add variadic assert command
Add variadic assert command such that all of the following are passing assertions

	assert true
	assert 1==1
	assert 1 1
	assert "abc" "abc" "comparing words"

Implement it as builtin in the evaluator package so that it can be used via an
`evy test` command in a follow up commit.

Add test reporting at the end of the program evaluation, if TestInfo.Reporting
isn't set to NoReporting and there are calls to assert.

Document it all.

Add examples/human-eval as initial conformance test suite to the Makefile
under make target `conform`.